Reorganizing uavcan message folder structure to be consistent with repository public_regulated_data_types
