+++ src/uavcan/node/Heartbeat.1.0.cpp
